requirements.txt
# Data Scripts
requests>=2.32.4
pandas>=2.3.0
opencv-python

# Birdiary viewer for data verification
pywebview>=5.4
pyinstaller>=6.14.1